Level Changes & Fixes
Reworked level lighting
Paddle blueprint now has a scene component as the root, allowing for a rotated capsule collider that fits around the mesh.

TODO: Tweak the paddle movement slightly so that it may go off screen, but loops back around to the opposite side of the viewport based off the side the paddle exits.